Move condition source files to Globals.
