+++ src/game/Globals/Conditions.h
